backend/Api/requirements.txt
torch~=1.13.1
Flask~=2.2.5
requests~=2.31.0
Pillow~=9.5.0
numpy~=1.21.6
opencv-python~=4.8.0.74
scikit-image~=0.19.3
matplotlib~=3.5.3
torchvision~=0.14.1
SQLAlchemy~=2.0.19